+++ src/lib.rs
@@ -7,6 +7,11 @@
 //! [`PresenceRunner`] - core runner for rich presence<br>
 //! [`PresenceClient`] - client/messenger instance used for sending activity and closing updates
 //!
+//! ## Looking for Discord's documentation?
+//!
+//! The groundwork of this project has been done using the official [Discord RPC documentation](https://docs.discord.com/developers/topics/rpc)
+//! page. Do note that not all features are/will be implemented as most of them are not fully implemented in Discord itself on a client basis.
+//!
 //! ## Examples
 //!
 //! Library examples can be found in [this directory](https://github.com/hitblast/filthy-rich/tree/master/examples) on GitHub.